Better update external-ui

+++ config/update_ui.go
@@ -4,20 +4,17 @@ import (
 	"archive/zip"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
 	"strings"
 	"sync"
-
-	C "github.com/Dreamacro/clash/constant"
 )
 
 var (
 	ExternalUIURL    string
 	ExternalUIPath   string
-	ExternalUIFolder string
-	ExternalUIName   string
 )
 
 var xdMutex sync.Mutex
@@ -26,11 +23,11 @@ func UpdateUI() error {
 	xdMutex.Lock()
 	defer xdMutex.Unlock()
 
-	if ExternalUIPath == "" || ExternalUIFolder == "" {
+	if ExternalUIPath == "" || ExternalUIURL == "" {
 		return fmt.Errorf("ExternalUI configure incomplete")
 	}
 
-	err := cleanup(ExternalUIFolder)
+	err := cleanup(ExternalUIPath)
 	if err != nil {
 		if !os.IsNotExist(err) {
 			return fmt.Errorf("cleanup exist file error: %w", err)
@@ -42,21 +39,29 @@ func UpdateUI() error {
 		return fmt.Errorf("can't download  file: %w", err)
 	}
 
-	saved := path.Join(C.Path.HomeDir(), "download.zip")
+	saved := path.Join(ExternalUIPath, "download.zip")
 	if saveFile(data, saved) != nil {
 		return fmt.Errorf("can't save zip file: %w", err)
 	}
 	defer os.Remove(saved)
 
-	unzipFolder, err := unzip(saved, C.Path.HomeDir())
+	unzipFolder, err := unzip(saved, ExternalUIPath)
 	if err != nil {
 		return fmt.Errorf("can't extract zip file: %w", err)
 	}
 
-	err = os.Rename(unzipFolder, ExternalUIFolder)
+	files, err := ioutil.ReadDir(unzipFolder)
 	if err != nil {
-		return fmt.Errorf("can't rename folder: %w", err)
+		return fmt.Errorf("error reading source folder: %w", err)
 	}
+
+	for _, file := range files {
+		err = os.Rename(filepath.Join(unzipFolder, file.Name()), filepath.Join(ExternalUIPath, file.Name()))
+		if err != nil {
+			return nil
+		}
+	}
+	defer os.Remove(unzipFolder)
 	return nil
 }
 
@@ -107,6 +112,10 @@ func cleanup(root string) error {
 	return filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
 			return err
+		}
+		if path == root {
+			// skip root itself
+			return nil
 		}
 		if info.IsDir() {
 			if err := os.RemoveAll(path); err != nil {
